requirements.txt
asgiref==3.6.0
Django==3.2.11
django-cors-headers==3.4.0
djangorestframework==3.12.4
Pillow==8.1.2
pytz==2023.2
sqlparse==0.4.3
